requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
attrs==20.3.0
    # via -r requirements.in
beautifulsoup4==4.9.3
    # via -r requirements.in
click==7.1.2
    # via
    #   -r requirements.in
    #   nltk
joblib==1.0.1
    # via nltk
nltk==3.5
    # via -r requirements.in
regex==2020.11.13
    # via nltk
soupsieve==2.2
    # via beautifulsoup4
tqdm==4.58.0
    # via
    #   -r requirements.in
    #   nltk
